+++ src/main/java/generator/LottoNumberGenerator.java
@@ -1,7 +1,7 @@
 package generator;
 
 import java.util.List;
-import lotto.LottoNumber;
+import lotto.model.LottoNumber;
 
 public interface LottoNumberGenerator {
     List<LottoNumber> generate();
